requirements.txt
python-telegram-bot>=20.0
Telethon>=1.30.0
python-dotenv>=1.0.0
emoji>=2.0.0
